PRMDR-472 bulk_upload_metadata_handler: move business logic to service (#189)
* [PRMDR-472] Connect lambda handler to new service class, fix unit tests

* [PRMDR-472] Run formatter

* [PRMDR-472] Fixing unit tests

* [PRMDR-472] Migrate main routine to process_metadata() method in service class

* [PRMDR-472] Add unit test for new service class

* [PRMDR-472] Clear temp folder at the end of lambda process

* [PRMDR-472] Address PR comment (follow Arrange-Act-Assert pattern)

* [PRMDR-472] Address PR comment (move BulkUploadMetadataService() instantiation to a fixture in unit tests)

* [PRMDR-472] Address PR comment (move specific errors catching into service class)
