+++ lambdas/tests/unit/helpers/data/bulk_upload/test_data.py
@@ -46,7 +46,9 @@ patient_2_file_1 = MetadataFile(
     upload="04/10/2023",
 )
 
-patient_2 = StagingMetadata(nhs_number="1234567891", files=[patient_2_file_1], retries=0)
+patient_2 = StagingMetadata(
+    nhs_number="1234567891", files=[patient_2_file_1], retries=0
+)
 
 EXPECTED_PARSED_METADATA = [patient_1, patient_2]
